Split test data and helpers across separate files
In preparation for standalone operators
