+++ testing/catalog.go
@@ -10,7 +10,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 	"k8s.io/api/apps/v1beta2"
-	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -18,10 +17,6 @@ import (
 
 	"code.cloudfoundry.org/cf-operator/pkg/bosh/manifest"
 	"code.cloudfoundry.org/cf-operator/pkg/credsgen"
-	bdv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/boshdeployment/v1alpha1"
-	ejv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedjob/v1alpha1"
-	esv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedsecret/v1alpha1"
-	essv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedstatefulset/v1alpha1"
 	"code.cloudfoundry.org/cf-operator/pkg/kube/util"
 	"code.cloudfoundry.org/cf-operator/pkg/kube/util/config"
 	bm "code.cloudfoundry.org/cf-operator/testing/boshmanifest"
@@ -200,16 +195,6 @@ func (c *Catalog) DefaultConfigMap(name string) corev1.ConfigMap {
 	}
 }
 
-// DefaultBOSHDeployment a deployment CR
-func (c *Catalog) DefaultBOSHDeployment(name, manifestRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: bdv1.BOSHDeploymentSpec{
-			Manifest: bdv1.ResourceReference{Name: manifestRef, Type: bdv1.ConfigMapReference},
-		},
-	}
-}
-
 // InterpolateOpsConfigMap for ops interpolate configmap tests
 func (c *Catalog) InterpolateOpsConfigMap(name string) corev1.ConfigMap {
 	return corev1.ConfigMap{
@@ -258,19 +243,6 @@ func (c *Catalog) InterpolateOpsIncorrectSecret(name string) corev1.Secret {
 	}
 }
 
-// DefaultExtendedSecret for use in tests
-func (c *Catalog) DefaultExtendedSecret(name string) esv1.ExtendedSecret {
-	return esv1.ExtendedSecret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: esv1.ExtendedSecretSpec{
-			Type:       "password",
-			SecretName: "generated-secret",
-		},
-	}
-}
-
 // DefaultCA for use in tests
 func (c *Catalog) DefaultCA(name string, ca credsgen.Certificate) corev1.Secret {
 	return corev1.Secret{
@@ -278,67 +250,6 @@ func (c *Catalog) DefaultCA(name string, ca credsgen.Certificate) corev1.Secret 
 		Data: map[string][]byte{
 			"ca":     ca.Certificate,
 			"ca_key": ca.PrivateKey,
-		},
-	}
-}
-
-// DefaultExtendedStatefulSet for use in tests
-func (c *Catalog) DefaultExtendedStatefulSet(name string) essv1.ExtendedStatefulSet {
-	return essv1.ExtendedStatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: essv1.ExtendedStatefulSetSpec{
-			Template: c.DefaultStatefulSet(name),
-		},
-	}
-}
-
-// WrongExtendedStatefulSet for use in tests
-func (c *Catalog) WrongExtendedStatefulSet(name string) essv1.ExtendedStatefulSet {
-	return essv1.ExtendedStatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: essv1.ExtendedStatefulSetSpec{
-			Template: c.WrongStatefulSet(name),
-		},
-	}
-}
-
-// ExtendedStatefulSetWithPVC for use in tests
-func (c *Catalog) ExtendedStatefulSetWithPVC(name, pvcName string, storageClassName string) essv1.ExtendedStatefulSet {
-	return essv1.ExtendedStatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: essv1.ExtendedStatefulSetSpec{
-			Template: c.StatefulSetWithPVC(name, pvcName, storageClassName),
-		},
-	}
-}
-
-// WrongExtendedStatefulSetWithPVC for use in tests
-func (c *Catalog) WrongExtendedStatefulSetWithPVC(name, pvcName string, storageClassName string) essv1.ExtendedStatefulSet {
-	return essv1.ExtendedStatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: essv1.ExtendedStatefulSetSpec{
-			Template: c.WrongStatefulSetWithPVC(name, pvcName, storageClassName),
-		},
-	}
-}
-
-// OwnedReferencesExtendedStatefulSet for use in tests
-func (c *Catalog) OwnedReferencesExtendedStatefulSet(name string) essv1.ExtendedStatefulSet {
-	return essv1.ExtendedStatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: essv1.ExtendedStatefulSetSpec{
-			UpdateOnConfigChange: true,
-			Template:             c.OwnedReferencesStatefulSet(name),
 		},
 	}
 }
@@ -660,118 +571,6 @@ func (c *Catalog) OwnedReferencesPodTemplate(name string) corev1.PodTemplateSpec
 	}
 }
 
-// CmdPodTemplate returns the spec with a given command for busybox
-func (c *Catalog) CmdPodTemplate(cmd []string) corev1.PodTemplateSpec {
-	return corev1.PodTemplateSpec{
-		Spec: corev1.PodSpec{
-			RestartPolicy:                 corev1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: util.Int64(1),
-			Containers: []corev1.Container{
-				{
-					Name:    "busybox",
-					Image:   "busybox",
-					Command: cmd,
-					Env: []corev1.EnvVar{
-						{Name: "REPLICAS", Value: "1"},
-						{Name: "AZ_INDEX", Value: "1"},
-						{Name: "POD_ORDINAL", Value: "0"},
-					},
-				},
-			},
-		},
-	}
-}
-
-// ConfigPodTemplate returns the spec with a given command for busybox
-func (c *Catalog) ConfigPodTemplate() corev1.PodTemplateSpec {
-	one := int64(1)
-	return corev1.PodTemplateSpec{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels: map[string]string{"delete": "pod"},
-		},
-		Spec: corev1.PodSpec{
-			RestartPolicy:                 corev1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: &one,
-			Volumes: []corev1.Volume{
-				{
-					Name: "secret1",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "secret1",
-						},
-					},
-				},
-				{
-					Name: "configmap1",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: "config1",
-							},
-						},
-					},
-				},
-			},
-			Containers: []corev1.Container{
-				{
-					Name:    "busybox",
-					Image:   "busybox",
-					Command: []string{"sleep", "1"},
-					Env: []corev1.EnvVar{
-						{Name: "REPLICAS", Value: "1"},
-						{Name: "AZ_INDEX", Value: "1"},
-						{Name: "POD_ORDINAL", Value: "0"},
-					},
-				},
-			},
-		},
-	}
-}
-
-// MultiContainerPodTemplate returns the spec with two containers running a given command for busybox
-func (c *Catalog) MultiContainerPodTemplate(cmd []string) corev1.PodTemplateSpec {
-	return corev1.PodTemplateSpec{
-		Spec: corev1.PodSpec{
-			RestartPolicy:                 corev1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: util.Int64(1),
-			Containers: []corev1.Container{
-				{
-					Name:    "busybox",
-					Image:   "busybox",
-					Command: cmd,
-				},
-				{
-					Name:    "busybox2",
-					Image:   "busybox",
-					Command: cmd,
-				},
-			},
-		},
-	}
-}
-
-// FailingMultiContainerPodTemplate returns a spec with a given command for busybox and a second container which fails
-func (c *Catalog) FailingMultiContainerPodTemplate(cmd []string) corev1.PodTemplateSpec {
-	return corev1.PodTemplateSpec{
-		Spec: corev1.PodSpec{
-			RestartPolicy:                 corev1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: util.Int64(1),
-			Containers: []corev1.Container{
-				{
-					Name:    "busybox",
-					Image:   "busybox",
-					Command: cmd,
-				},
-				{
-					Name:    "failing",
-					Image:   "busybox",
-					Command: []string{"exit", "1"},
-				},
-			},
-		},
-	}
-}
-
 // DefaultPod defines a pod with a simple web server useful for testing
 func (c *Catalog) DefaultPod(name string) corev1.Pod {
 	return corev1.Pod{
@@ -814,131 +613,6 @@ func (c *Catalog) Sleep1hPodSpec() corev1.PodSpec {
 				Image:   "busybox",
 				Command: []string{"sleep", "3600"},
 			},
-		},
-	}
-}
-
-// EmptyBOSHDeployment empty deployment CR
-func (c *Catalog) EmptyBOSHDeployment(name, manifestRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec:       bdv1.BOSHDeploymentSpec{},
-	}
-}
-
-// DefaultBOSHDeploymentWithOps a deployment CR with ops
-func (c *Catalog) DefaultBOSHDeploymentWithOps(name, manifestRef string, opsRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: bdv1.BOSHDeploymentSpec{
-			Manifest: bdv1.ResourceReference{Name: manifestRef, Type: bdv1.ConfigMapReference},
-			Ops: []bdv1.ResourceReference{
-				{Name: opsRef, Type: bdv1.ConfigMapReference},
-			},
-		},
-	}
-}
-
-// WrongTypeBOSHDeployment a deployment CR containing wrong type
-func (c *Catalog) WrongTypeBOSHDeployment(name, manifestRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: bdv1.BOSHDeploymentSpec{
-			Manifest: bdv1.ResourceReference{Name: manifestRef, Type: "wrong-type"},
-		},
-	}
-}
-
-// BOSHDeploymentWithWrongTypeOps a deployment CR with wrong type ops
-func (c *Catalog) BOSHDeploymentWithWrongTypeOps(name, manifestRef string, opsRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: bdv1.BOSHDeploymentSpec{
-			Manifest: bdv1.ResourceReference{Name: manifestRef, Type: bdv1.ConfigMapReference},
-			Ops: []bdv1.ResourceReference{
-				{Name: opsRef, Type: "wrong-type"},
-			},
-		},
-	}
-}
-
-// InterpolateBOSHDeployment a deployment CR
-func (c *Catalog) InterpolateBOSHDeployment(name, manifestRef, opsRef string, secretRef string) bdv1.BOSHDeployment {
-	return bdv1.BOSHDeployment{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: bdv1.BOSHDeploymentSpec{
-			Manifest: bdv1.ResourceReference{Name: manifestRef, Type: bdv1.ConfigMapReference},
-			Ops: []bdv1.ResourceReference{
-				{Name: opsRef, Type: bdv1.ConfigMapReference},
-				{Name: secretRef, Type: bdv1.SecretReference},
-			},
-		},
-	}
-}
-
-// DefaultExtendedJob default values
-func (c *Catalog) DefaultExtendedJob(name string) *ejv1.ExtendedJob {
-	cmd := []string{"sleep", "1"}
-	return &ejv1.ExtendedJob{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: ejv1.ExtendedJobSpec{
-			Trigger: ejv1.Trigger{
-				Strategy: ejv1.TriggerNow,
-			},
-			Template: c.CmdPodTemplate(cmd),
-		},
-	}
-}
-
-// DefaultExtendedJobWithSucceededJob returns an ExtendedJob and a Job owned by it
-func (c *Catalog) DefaultExtendedJobWithSucceededJob(name string) (*ejv1.ExtendedJob, *batchv1.Job, *corev1.Pod) {
-	ejob := c.DefaultExtendedJob(name)
-	backoffLimit := util.Int32(2)
-	job := &batchv1.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name + "-job",
-			OwnerReferences: []metav1.OwnerReference{
-				{
-					Name:       name,
-					UID:        "",
-					Controller: util.Bool(true),
-				},
-			},
-		},
-		Spec:   batchv1.JobSpec{BackoffLimit: backoffLimit},
-		Status: batchv1.JobStatus{Succeeded: 1},
-	}
-	pod := c.DefaultPod(name + "-pod")
-	pod.Labels = map[string]string{
-		"job-name": job.GetName(),
-	}
-	return ejob, job, &pod
-}
-
-// ErrandExtendedJob default values
-func (c *Catalog) ErrandExtendedJob(name string) ejv1.ExtendedJob {
-	cmd := []string{"sleep", "1"}
-	return ejv1.ExtendedJob{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: ejv1.ExtendedJobSpec{
-			Trigger: ejv1.Trigger{
-				Strategy: ejv1.TriggerNow,
-			},
-			Template: c.CmdPodTemplate(cmd),
-		},
-	}
-}
-
-// AutoErrandExtendedJob default values
-func (c *Catalog) AutoErrandExtendedJob(name string) ejv1.ExtendedJob {
-	cmd := []string{"sleep", "1"}
-	return ejv1.ExtendedJob{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: ejv1.ExtendedJobSpec{
-			Trigger: ejv1.Trigger{
-				Strategy: ejv1.TriggerOnce,
-			},
-			Template: c.CmdPodTemplate(cmd),
 		},
 	}
 }
